Refactored code. Removed debug print statements to /tmp log file

+++ ssl/ssl_sess.c
@@ -916,11 +916,11 @@ long SSL_SESSION_get_time(const SSL_SESSION *s)
     return (long)ossl_time_to_time_t(s->time);
 }
 
-uint64_t SSL_SESSION_get_rtt(const SSL_SESSION *s) // return rtt in microseconds. Matches the units of nginx's $tcpinfo_rtt
+long SSL_SESSION_get_rtt(const SSL_SESSION *s) // return rtt in microseconds. Matches the units of nginx's $tcpinfo_rtt
 {
     if (s == NULL)
         return 0;
-    return ossl_time2us(s->rtt);
+    return (long)ossl_time2us(s->rtt);
 }
 
 long SSL_SESSION_set_time(SSL_SESSION *s, long t)
